pyproject.toml
[tool.poetry]
name = "trainer3"
version = "0.1.0"
description = ""
authors = ["CTOaerofoyl <cto@aerofoyl.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"
ultralytics = "^8.3.38"
torch = {version = "^2.5.1+cu118", source = "pytorch"}
torchvision = {version = "^0.20.1+cu118", source = "pytorch"}
torchaudio = {version = "^2.5.1+cu118", source = "pytorch"}
tabulate = "^0.9.0"
PyQt5-sip = "^12.15.0"
qtawesome = "^1.3.1"
opencv-python = "^4.10.0.84"
scikit-image = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"